Added port forwarding and fixed Chef barbican api recipe
The generated Barbican API node was functional, but could not be reached from outside
until the port forwarding (for 9311 and 9312) was added.

The Chef api recipe was also not idempotent regarding the Fedora EPEL installation step,
so that was fixed to first check to see if that rpm was already installed on the node.